Move HTTP tests to submodule
